% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaDEA.R
\name{limmaDEA}
\alias{limmaDEA}
\title{Robust Wrapper for Differential Analysis with LIMMA}
\usage{
limmaDEA(
  fit,
  coef = NULL,
  test = c("t", "F"),
  adjust.method = "BH",
  adjust.globally = FALSE
)
}
\arguments{
\item{fit}{an \code{\link[limma:MArrayLM-class]{MArrayLM}} object. For
example, the output of \code{\link{limmaFit}}.}

\item{coef}{indices or names of columns of \code{fit$coefficients} to test.}

\item{test}{character; the type of statistical test to perform for each
feature. Either \code{"t"} (moderated t-tests) or \code{"F"} (moderated
F-tests).}

\item{adjust.method}{character; one of \code{p.adjust.methods}. Default is
\code{"BH"}.}

\item{adjust.globally}{logical; whether to perform global p-value adjustment
for sets of related contrasts.}
}
\value{
Output of \code{\link[limma]{topTable}} with additional columns
\code{"contrast"} and \code{"df.total"}.
}
\description{
A wrapper function for \code{\link[limma]{topTable}}. Especially
useful when performing moderated t-tests for more than one coefficient or
contrast.
}
\details{
See \code{\link{limmaFit}} to easily create
\code{\link[limma:MArrayLM-class]{MArrayLM}} objects for \code{fit}.

If \code{adjust.globally=TRUE}, the entire vector of p-values will be
adjusted together. Only do this if sets of related contrasts are being
tested. Otherwise, p-values are adjusted separately by contrast using
\code{adjust.method}. By default \code{adjust.method = "BH"}, which uses the
method of Benjamini & Hochberg (1995) to control the false discovery rate.
}
\references{
Benjamini, Y., and Hochberg, Y. (1995). Controlling the false discovery rate:
a practical and powerful approach to multiple testing. \emph{Journal of the
Royal Statistical Society Series B}, \strong{57}, 289–300.
doi:\href{https://doi.org/10.1111/j.2517-6161.1995.tb02031.x}{10.1111/j.2517-6161.1995.tb02031.x}.

Ritchie ME, Phipson B, Wu D, Hu Y, Law CW, Shi W, and Smyth GK (2015). limma
powers differential expression analyses for RNA-sequencing and microarray
studies. \emph{Nucleic Acids Research} 43, e47.
\url{http://nar.oxfordjournals.org/content/43/7/e47}
}
\seealso{
\code{\link[limma]{limmaUsersGuide}}, \code{\link{limmaFit}},
\code{\link[limma]{topTable}}
}
