% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaFit.R
\name{limmaFit}
\alias{limmaFit}
\title{Robust Wrapper for Linear Modeling with LIMMA}
\usage{
limmaFit(
  object,
  model.str,
  contrasts,
  trend = TRUE,
  robust = TRUE,
  var.group = NULL,
  block = NULL,
  plot = FALSE
)
}
\arguments{
\item{object}{object inheriting from class \code{\link[Biobase]{eSet}}.}

\item{model.str}{character; model formula of the form \code{"~ a + b"} or
\code{"~ 0 + a + b"} (no intercept), where \code{"a"}, and \code{"b"} are
predictors (usually columns of \code{pData(object)}.}

\item{contrasts}{character; (optional) one or more contrasts of the form
\code{"varA - varB"}, where \code{var} is a predictor included in
\code{model.str}. Contrasts involving interactions are not allowed. See
See \code{paircomp} for assistance with creating these contrasts.}

\item{trend}{logical, character; should an intensity-dependent trend be
allowed for the prior variance? If \code{FALSE}, then the prior variance is
constant. Alternatively, \code{trend} can be the name of a column in
\code{fData(object)}, which will be used as the covariate for the prior
variance. Passed to \code{\link[limma]{eBayes}}.}

\item{robust}{logical; should the estimation of \code{df.prior} and
\code{var.prior} be made robust to hyper- or hypo-variable features? Passed
to \code{\link[limma]{eBayes}}.}

\item{var.group}{character; name of a column in \code{pData(object)} that
groups samples for \code{\link[limma]{arrayWeights}}. If \code{NULL}
(default), samples will be weighted equally.}

\item{block}{character; name of a column in \code{pData(object)} specifying a
blocking factor used in \code{\link[limma]{duplicateCorrelation}}. If
\code{NULL} (default), \code{duplicateCorrelation} is not used.}

\item{plot}{logical; whether to show model summary/diagnostic plots.}
}
\value{
An object of class \code{\link[limma:MArrayLM-class]{MArrayLM}}.
}
\description{
A \acronym{LIMMA} wrapper for -omics data in the form of
\code{\link[Biobase]{eSet}} or similar objects, including a separate
pipeline for RNA-Seq count data.
}
\section{Model Specification}{


If at least one predictor in \code{model.str} is categorical, exclusion of
the intercept merely changes the parameterization of the model from an
\strong{effects} model to a \strong{means} model.

\strong{Effects} model:

\deqn{y_{ij} = \mu_i + \epsilon_{ij}}

\strong{Means} model:

\deqn{y_{ij} = \mu + \tau_i + \epsilon_{ij}}

where \eqn{y_{ij}} is the \eqn{ij}th observation, \eqn{\mu_i} is the mean
of the \eqn{i}th level of the categorical predictor, \eqn{\mu} is the
overall mean, \eqn{\tau_i} is the \eqn{i}th treatment effect (difference
between the mean of the \eqn{i}th level and the overall mean), and
\eqn{\epsilon_{ij}} is the random error.

A no-intercept model is required if \code{contrasts} are provided.

If \code{model.str} is a no-intercept model with all numeric predictors, an
intercept will be recommended with a warning, since this makes the (often
incorrect) assumption that the regression line should pass through the
origin.
}

\section{General Details}{


If \code{trend} is the name of a column in \code{fData(object)}, missing
values are not allowed.

If there are fewer than 10 features, a warning will be issued, both
\code{robust} and \code{trend} will be set to \code{FALSE}, and samples
will be weighted equally.
}

\section{Modeling (RNA-Seq) Count Data}{


If \code{exprs(object)} is a matrix of counts, the voom pipeline for
RNA-Seq count data will be used. In that case, \code{trend} will be set to
\code{FALSE}, since \code{\link[edgeR]{voomLmFit}} incorporates the
mean-variance trend into the precision weights. Low-count features will be
removed (with a warning) by \code{\link[edgeR]{filterByExpr}}.

If there are fewer than 10 features, a warning will be issued, \code{trend}
will be set to \code{FALSE}, and samples will be weighted equally.
}

\examples{
library(Biobase)

# Simulate ExpressionSet object
set.seed(0)
nsamples <- 12L
nfeatures <- 1000L
mat_dimnames <- list(paste0("feature.", seq_len(nfeatures)),
                     paste0("sample.", seq_len(nsamples)))
mat <- matrix(rnorm(nsamples * nfeatures), ncol = nsamples,
              dimnames = mat_dimnames)

p_data <- data.frame(group = rep(LETTERS[1:2], each = nsamples / 2),
                     covariate = rpois(n = nsamples, lambda = 3),
                     row.names = colnames(mat))
p_data <- AnnotatedDataFrame(p_data)

obj <- ExpressionSet(assayData = mat, phenoData = p_data)

# Intercept model, categorical predictor
fit1 <- limmaFit(obj, model.str = "~ group")
head(fit1$coefficients)

# No-intercept model, categorical predictor
fit2 <- limmaFit(obj, model.str = "~ 0 + group")
head(fit2$coefficients)

# Intercept model, categorical and numeric predictors
fit3 <- limmaFit(obj, model.str = "~ group + covariate")
head(fit3$coefficients)

# No-intercept model, categorical and numeric predictors
fit4 <- limmaFit(obj, model.str = "~ 0 + group + covariate")
head(fit4$coefficients)

# Include contrasts (requires no-intercept model)
fit5 <- limmaFit(obj, model.str = "~ 0 + group",
                 contrasts = "groupA - groupB")
head(fit5$coefficients) # coefficients same as fit5$coefficients[, "groupB"]

}
\references{
Anders, S, Huber, W (2010). Differential expression analysis for sequence
count data. \emph{Genome Biology} 11, R106.

Bullard JH, Purdom E, Hansen KD, Dudoit S. (2010) Evaluation of statistical
methods for normalization and differential expression in mRNA-Seq
experiments. \emph{BMC Bioinformatics} 11, 94.

Chen Y, Lun ATL, and Smyth, GK (2016). From reads to genes to pathways:
differential expression analysis of RNA-Seq experiments using Rsubread and
the edgeR quasi-likelihood pipeline. \emph{F1000Research} 5, 1438.
\url{https://f1000research.com/articles/5-1438}

Law, CW, Chen, Y, Shi, W, Smyth, GK (2014). Voom: precision weights unlock
linear model analysis tools for RNA-seq read counts. \emph{Genome Biology}
15, R29. See also the Preprint Version at
\url{http://www.statsci.org/smyth/pubs/VoomPreprint.pdf} incorporating some
notational corrections.

Lun, ATL, Smyth, GK (2017). No counts, no variance: allowing for loss of
degrees of freedom when assessing biological variability from RNA-seq data.
Statistical Applications in Genetics and Molecular Biology 16(2), 83-93.
\href{https://www.doi.org/10.1515/sagmb-2017-0010}{doi:10.1515/sagmb-2017-0010}

Liu, R, Holik, AZ, Su, S, Jansz, N, Chen, K, Leong, HS, Blewitt, ME,
Asselin-Labat, M-L, Smyth, GK, Ritchie, ME (2015). Why weight? Modelling
sample and observational level variability improves power in RNA-seq
analyses. \emph{Nucleic Acids Research 43}, e97.
\href{https://www.doi.org/10.1093/nar/gkv412}{doi:10.1093/nar/gkv412}

Phipson, B, Lee, S, Majewski, IJ, Alexander, WS, and Smyth, GK (2016). Robust
hyperparameter estimation protects against hypervariable genes and improves
power to detect differential expression. \emph{Annals of Applied Statistics}
10, 946-963. \url{http://projecteuclid.org/euclid.aoas/1469199900}

Ritchie, M. E., Diyagama, D., Neilson, van Laar, R., J., Dobrovic, A.,
Holloway, A., and Smyth, G. K. (2006). Empirical array quality weights in the
analysis of microarray data. \emph{BMC Bioinformatics} \strong{7}, 261.
\href{https://doi.org/10.1186/1471-2105-7-261}{doi:10.1186/1471-2105-7-261}

Ritchie ME, Phipson B, Wu D, Hu Y, Law CW, Shi W, and Smyth GK (2015). limma
powers differential expression analyses for RNA-sequencing and microarray
studies. \emph{Nucleic Acids Research} 43, e47.
\url{http://nar.oxfordjournals.org/content/43/7/e47}

Robinson MD, Oshlack A (2010). A scaling normalization method for
differential expression analysis of RNA-seq data. \emph{Genome Biology} 11,
R25.

Smyth, G. K. (2002). An efficient algorithm for REML in heteroscedastic
regression. \emph{Journal of Computational and Graphical Statistics}
\strong{11}, 836-847. \url{https://gksmyth.github.io/pubs/remlalgo.pdf}

Smyth, G. K. (2004). Linear models and empirical Bayes methods for assessing
differential expression in microarray experiments. \emph{Statistical
Applications in Genetics and Molecular Biology}, \strong{3}, No. 1, Article
3. https://gksmyth.github.io/pubs/ebayes.pdf

Smyth, G. K., Michaud, J., and Scott, H. (2005). The use of within-array
replicate spots for assessing differential expression in microarray
experiments. \emph{Bioinformatics 21}(9), 2067-2075.
\url{http://bioinformatics.oxfordjournals.org/content/21/9/2067}
Preprint with corrections: \url{https://gksmyth.github.io/pubs/dupcor.pdf}
}
\seealso{
\code{\link[limma]{limmaUsersGuide}}, \code{\link[limma]{lmFit}},
}
