% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{molecular_signatures}
\alias{molecular_signatures}
\alias{human_gene_sets}
\alias{rat_gene_sets}
\alias{SET_TO_ID}
\alias{acetyl_GN_NORM_gene_sets}
\alias{acetyl_GN_gene_sets}
\alias{human_kinase_sets}
\alias{metab_GN_refmet_sets}
\alias{metab_VL_refmet_sets}
\alias{phospho_GN_NORM_kinase_sets}
\alias{phospho_GN_kinase_sets}
\alias{prot_GN_gene_sets}
\alias{redox_GN_NORM_gene_sets}
\alias{redox_GN_gene_sets}
\alias{trnscrpt_GN_gene_sets}
\alias{trnscrpt_VL_gene_sets}
\title{Molecular Signatures}
\format{
Named lists of length 10472 (\code{human_gene_sets}), 9293
(\code{rat_gene_sets}), or 319 (\code{human_kinase_sets}).
}
\usage{
human_gene_sets
rat_gene_sets
human_kinase_sets

## Dataset-specific sets ----
# Gastrocnemius
trnscrpt_GN_gene_sets  # Transcriptomics
prot_GN_gene_sets      # Proteomics
phospho_GN_kinase_sets # Phosphoproteomics
acetyl_GN_gene_sets    # Acetyl proteomics
redox_GN_gene_sets     # Redox proteomics
metab_GN_refmet_sets   # Metabolomics

# Vastus lateralis
trnscrpt_VL_gene_sets  # Transcriptomics
#'
}
\description{
Gene sets from the C5 and C8 collections of the Molecular
Signatures Database (MSigDB) and phosphorylation sites grouped by their
known kinases from PhosphoSitePlus. Used to summarize results of
differential analysis performed on -omics datasets.
}
\details{
Human gene sets were obtained from version 2023.2.Hs of the
Molecular Signatures Database (MSigDB): specifically, those gene sets from
the C5:GO subcollection (Gene Ontology terms) and the
"RUBENSTEIN_SKELETAL_MUSCLE" sets from the C8 collection (Ashburner, 2000;
Liberzon 2011, 2015; Rubenstein, 2020; The Gene Ontology Consortium, 2023).

Human genes were mapped to their rat orthologs using
\code{\link[babelgene]{orthologs}} (v22.9; \code{\link{RAT_TO_HUMAN}}).
Only those rat gene sets with at least 5 genes were retained.

Kinase–Substrate relationship data was obtained from
"Kinase_Substrate_Dataset.gz", downloaded from PhosphoSitePlus
(\acronym{PSP}) v6.7.1.1 (last modified Fri Nov 17 08:50:20 EST 2023):
\url{https://www.phosphosite.org/staticDownloads.action}. Only human
kinases and substrates were retained, since the data is substantially more
comprehensive for humans than any other organism. Furthermore, only those
kinases with at least 3 known substrates were kept. This results in
phosphorylation sites grouped into 319 kinases.

Metabolomics RefMet subclass information was downloaded from
\url{https://www.metabolomicsworkbench.org/databases/refmet/name_to_refmetF_form.php}
(accessed 2024-23-03).

Dataset-specific molecular signatures were generated by restricting each
set to only those features measured in a particular tissue/ome combination
with \code{\link[TMSig]{filterSets}} and only keeping those sets with at
least 5 genes or metabolites/lipids or 3 human phosphorylation sites.
}
\note{
PhosphoSitePlus data is not for commercial use.
}
\references{
Ashburner, M., Ball, C. A., Blake, J. A., Botstein, D., Butler,
H., Cherry, J. M., Davis, A. P., Dolinski, K., Dwight, S. S., Eppig, J. T.,
Harris, M. A., Hill, D. P., Issel-Tarver, L., Kasarskis, A., Lewis, S.,
Matese, J. C., Richardson, J. E., Ringwald, M., Rubin, G. M., & Sherlock,
G. (2000). Gene Ontology: Tool for the unification of biology. \emph{Nature
Genetics, 25}(1), 25–29. \url{https://doi.org/10.1038/75556}

Liberzon, A., Subramanian, A., Pinchback, R., Thorvaldsdóttir, H., Tamayo,
P., & Mesirov, J. P. (2011). Molecular signatures database (MSigDB) 3.0.
\emph{Bioinformatics, 27}(12), 1739–1740.
\url{https://doi.org/10.1093/bioinformatics/btr260}

Hornbeck, P. V., Zhang, B., Murray, B., Kornhauser, J. M., Latham, V., &
Skrzypek, E. (2015). PhosphoSitePlus, 2014: mutations, PTMs and
recalibrations. \emph{Nucleic acids research, 43}, D512–D520.
\url{https://doi.org/10.1093/nar/gku1267}

Liberzon, A., Birger, C., Thorvaldsdóttir, H., Ghandi, M., Mesirov, J. P.,
& Tamayo, P. (2015). The Molecular Signatures Database (MSigDB) hallmark
gene set collection. \emph{Cell systems, 1}(6), 417–425.
\url{https://doi.org/10.1016/j.cels.2015.12.004}

Rubenstein, A. B., Smith, G. R., Raue, U., Begue, G., Minchev, K.,
Ruf-Zamojski, F., Nair, V. D., Wang, X., Zhou, L., Zaslavsky, E., Trappe,
T. A., Trappe, S., & Sealfon, S. C. (2020). Single-cell transcriptional
profiles in human skeletal muscle. \emph{Scientific reports, 10}(1), 229.
\url{https://doi.org/10.1038/s41598-019-57110-6}

Martin, F. J., Amode, M. R., Aneja, A., Austine-Orimoloye, O., Azov, A. G.,
Barnes, I., Becker, A., Bennett, R., Berry, A., Bhai, J., Bhurji, S. K.,
Bignell, A., Boddu, S., Branco Lins, P. R., Brooks, L., Ramaraju, S. B.,
Charkhchi, M., Cockburn, A., Da Rin Fiorretto, L., … Flicek, P. (2023).
Ensembl 2023. \emph{Nucleic Acids Research, 51}(D1), D933–D941.
\url{https://doi.org/10.1093/nar/gkac958}

The Gene Ontology Consortium, Aleksander, S. A., Balhoff, J., Carbon, S.,
Cherry, J. M., Drabkin, H. J., Ebert, D., Feuermann, M., Gaudet, P.,
Harris, N. L., Hill, D. P., Lee, R., Mi, H., Moxon, S., Mungall, C. J.,
Muruganugan, A., Mushayahama, T., Sternberg, P. W., Thomas, P. D., …
Westerfield, M. (2023). The Gene Ontology knowledgebase in 2023.
\emph{Genetics, 224}(1), iyad031.
\url{https://doi.org/10.1093/genetics/iyad031}
}
\keyword{datasets}
