% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscle_ora.R
\name{muscle_ora}
\alias{muscle_ora}
\title{Overrepresentation analysis for specific gene sets}
\usage{
muscle_ora(input, background, database = rat_gene_sets, min_size = 5L)
}
\arguments{
\item{input}{the set of features you would like to investigate}

\item{background}{the set of all features included/studied}

\item{database}{the gene sets you would like to include. See \code{rat_gene_sets} for information about the specific format that is expected.}

\item{min_size}{Minimum number of features in a gene set included in the \code{background} object}
}
\value{
a data.frame object that details enrichment results for each of the sets describes in \code{rat_gene_sets}
or whatever is used for the \code{database} input.
}
\description{
Overrepresentation analysis for specific gene sets
}
