% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{normalize_data}
\alias{normalize_data}
\title{Median-MAD-normalize expression data}
\usage{
normalize_data(x, mad = FALSE)
}
\arguments{
\item{x}{a matrix with features (genes, proteins, etc.) as rows and samples
as columns.}

\item{mad}{logical; whether to divide the median-centered samples by the mean
absolute deviations (MADs) and then multiply by the mean of the MADS.
Default \code{FALSE} only median-centers the samples.}
}
\value{
output of \code{\link[base]{scale}}. The centered, optionally scaled
matrix.
}
\description{
Normalize a matrix of log\eqn{_2}-transformed expression data.
Median-center columns and optionally scale by the mean absolute deviations
(MADs).
}
\examples{
set.seed(0)
x <- matrix(rnorm(200, sd = 2), nrow = 20, ncol = 10)
boxplot(x) # before normalization

x1 <- normalize_data(x, mad = FALSE)
boxplot(x1) # median-center samples only

x2 <- normalize_data(x, mad = TRUE)
boxplot(x2) # median-MAD normalize

# Sample medians and MADs
attr(x2, "scaled:center") # medians
attr(x2, "scaled:scale") # mads

}
