% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RAT_TO_HUMAN}
\alias{RAT_TO_HUMAN}
\title{Rat-to-human gene conversion table}
\format{
An object of class \code{data.frame} with 13023 rows and 9 columns.
}
\usage{
RAT_TO_HUMAN
}
\description{
Output of \code{\link[babelgene]{orthologs}} (babelgene v22.9).
}
\examples{
head(RAT_TO_HUMAN)

}
\keyword{datasets}
