% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RAT_TO_HUMAN_SITE}
\alias{RAT_TO_HUMAN_SITE}
\title{Rat-to-human gene conversion table for phophosites}
\format{
An object of class \code{data.frame} with 21570 rows and 2 columns.
}
\usage{
RAT_TO_HUMAN_SITE
}
\description{
Output of \code{\link[babelgene]{orthologs}} (babelgene v22.9).
}
\keyword{datasets}
