% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SKM_DA}
\alias{SKM_DA}
\alias{TRNSCRPT_GN_DA}
\alias{PROT_GN_DA}
\alias{PHOSPHO_GN_DA}
\alias{ACETYL_GN_DA}
\alias{REDOX_GN_DA}
\alias{METAB_GN_DA}
\alias{TRNSCRPT_VL_DA}
\alias{METAB_VL_DA}
\title{Skeletal muscle differential analysis results}
\format{
A named list of 2 \code{data.frame} objects:

\describe{
\item{"trained_vs_SED"}{All trained timepoints compared to their
sex-matched sedentary control group. Example: F_1W - F_SED. Total of 8
contrasts.}
\item{"MvF"}{Males compared to females at SED and 8W timepoints.}
}

In addition to all columns present in the \code{Biobase::fData} tables,
each \code{data.frame} contains the following variables, most of which
are added by \code{\link[limma:topTable]{limma::topTable}}:

\describe{
\item{\code{contrast}}{factor; the contrast(s) being tested.}
\item{\code{featureName}}{character; uniquely defines each feature that
was tested.}
\item{\code{logFC}}{numeric; difference in the weighted mean log\eqn{_2}
values of the groups specified in the \code{contrast} column.}
\item{\code{AveExpr}}{numeric; mean log\eqn{_2} relative abundance of
each feature.}
\item{\code{t}}{numeric; \acronym{LIMMA} moderated t-statistic.}
\item{\code{z}}{numeric; Z-score.}
\item{\code{P.Value}}{numeric; two-tailed p-value.}
\item{\code{adj.P.Val}}{numeric; Benjamini-Hochberg adjusted p-value.
P-values are adjusted across all trained vs. SED or male vs. female
contrasts.}
\item{\code{df.total}}{numeric; sum of residual and prior degrees of
freedom.}
\item{\code{signif_protein}}{logical; (PTM results only) whether the
corresponding protein was differentially expressed.}
}
}
\usage{
# Gastrocnemius
TRNSCRPT_GN_DA # Transcriptomics
PROT_GN_DA     # Proteomics
PHOSPHO_GN_DA  # Phosphoproteomics
ACETYL_GN_DA   # Acetyl proteomics
REDOX_GN_DA    # Redox proteomics
METAB_GN_DA    # Metabolomics

# Vastus lateralis
TRNSCRPT_VL_DA # Transcriptomics
METAB_VL_DA    # Metabolomics
}
\description{
Differential analysis of transcripts, proteins, phosphorylation
sites, acetylation sites, redox sites, and metabolites measured in rat
skeletal muscle (gastrocnemius and vastus lateralis).
}
\details{
Code to reproduce the analyses is provided in the
"differential_analysis" vignette.
}
\examples{
# Number of differential features (FDR = 0.05)
f <- function(x) {
  lapply(x, function(.x) with(.x, table(contrast, adj.P.Val < 0.05)))
}

## Gastrocnemius
f(TRNSCRPT_GN_DA) # Transcripts
f(PROT_GN_DA)     # Proteins
f(PHOSPHO_GN_DA)  # Phosphosites
f(ACETYL_GN_DA)   # Acetylation sites
f(REDOX_GN_DA)    # Redox sites
f(METAB_GN_DA)    # Metabolites

## Vastus lateralis
f(TRNSCRPT_VL_DA) # Transcripts
f(METAB_VL_DA)    # Metabolites

}
\seealso{
\code{\link[limma]{topTable}}
}
\keyword{datasets}
