% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SKM_GST}
\alias{SKM_GST}
\alias{TRNSCRPT_GN_FGSEA}
\alias{TRNSCRPT_GN_CAMERA}
\alias{PROT_GN_FGSEA}
\alias{PROT_GN_CAMERA}
\alias{ACETYL_GN_FGSEA}
\alias{ACETYL_GN_CAMERA}
\alias{ACETYL_GN_NORM_FGSEA}
\alias{ACETYL_GN_NORM_CAMERA}
\alias{PHOSPHO_GN_FGSEA}
\alias{PHOSPHO_GN_CAMERA}
\alias{PHOSPHO_GN_NORM_FGSEA}
\alias{PHOSPHO_GN_NORM_CAMERA}
\alias{REDOX_GN_FGSEA}
\alias{REDOX_GN_CAMERA}
\alias{REDOX_GN_NORM_FGSEA}
\alias{REDOX_GN_NORM_CAMERA}
\alias{METAB_GN_FGSEA}
\alias{METAB_GN_CAMERA}
\alias{TRNSCRPT_VL_FGSEA}
\alias{TRNSCRPT_VL_CAMERA}
\alias{METAB_VL_FGSEA}
\alias{METAB_VL_CAMERA}
\title{Molecular Signature Analysis Results}
\format{
Named lists of 2 \code{data.frame} objects:

\describe{
\item{"trained_vs_SED"}{All trained timepoints compared to their
sex-matched sedentary control group. Example: F_1W - F_SED. Total of 8
contrasts.}
\item{"MvF"}{Males compared to females at SED and 8W timepoints.}
}

Each \code{data.frame} contains the following columns:

\describe{
\item{\code{"Contrast"}}{the comparison being made. Indicates which
subset of the differential analysis results were used for testing.}
\item{\code{"GeneSetID"}}{a unique identifier for each gene set (metabolomics and phosphoproteomics excluded). Used for cross-referencing and shortening descriptions for plotting.}
\item{\code{"GeneSet"}}{a description of the gene set that was tested (metabolomics and phosphoproteomics excluded).}
\item{\code{"GeneSetShort"}}{a shortened version of \code{"GeneSet"} used for plotting (metabolomics and phosphoproteomics only). If the description in \code{"GeneSet"} was above a certain length, the string is broken at the space between words and the unique \code{"GeneSetID"} is appended at the end.}
\item{\code{"Subclass"}}{Metabolomics/lipidomics results only. A
description of the RefMet chemical subclass set that was tested.}
\item{\code{"Kinase"}}{Phosphoproteomics results only. The kinase that was
tested.}
\item{\code{"NGenesRat"}}{number of rat genes in the DA results for that
contrast that match genes in \code{"GeneSet"}.}
\item{\code{"NMetabolites"}}{Metabolomics/lipidomics results only. The
number of RefMet metabolite/lipid identifiers from the DA results for that
contrast that match a chemical subclass in \code{"Subclass"}.}
\item{\code{"NSites"}}{Phosphoproteomics results only. The number of human
sites from the DA results for that contrast that match sites in
\code{"Kinase"}.}
\item{\code{"NGenesHuman"}}{the size of the human gene set in MSigDB,
before any filtering.}
\item{\code{"NRatio"}}{\code{NGenesRat} / \code{NGenesHuman}. A measure of
confidence that the set being tested is what is described by
\code{"GeneSet"}.}
\item{\code{"Direction"}}{direction of change of the gene set. Factor with
levels "Up", "Down".}
\item{\code{"NES"}}{FGSEA results only. The normalized enrichment score (NES). The true enrichment score (ES) divided by the mean of the permutation enrichment scores with the same sign.}
\item{\code{"ZScore"}}{Z-score.}
\item{\code{"PValue"}}{two-sided p-values.}
\item{\code{"PAdj"}}{the Benjamini–Hochberg-adjusted p-value. P-values are
adjusted across all gene sets and contrasts.}
\item{\code{"LeadingEdge"}}{FGSEA only. A list of genes that contributed to
the enrichment score. Usually, this is a subset of all members of the set.}
}
}
\usage{
## FGSEA ----
# Gastrocnemius
TRNSCRPT_GN_FGSEA # Transcriptomics
PROT_GN_FGSEA     # Proteomics
ACETYL_GN_FGSEA   # Acetyl proteomics
REDOX_GN_FGSEA    # Redox proteomics
METAB_GN_FGSEA    # Metabolomics/lipidomics

# Vastus lateralis
TRNSCRPT_VL_FGSEA # Transcriptomics
METAB_VL_FGSEA    # Metabolomics/lipidomics

## CAMERA-PR ----
# Gastrocnemius
TRNSCRPT_GN_CAMERA # Transcriptomics
PROT_GN_CAMERA     # Proteomics
ACETYL_GN_CAMERA   # Acetyl proteomics
REDOX_GN_CAMERA    # Redox proteomics
METAB_GN_CAMERA    # Metabolomics/lipidomics

# Vastus lateralis
TRNSCRPT_VL_CAMERA # Transcriptomics
METAB_VL_CAMERA    # Metabolomics/lipidomics
}
\description{
Fast gene set enrichment analysis (FGSEA) and pre-ranked
Correlation Adjusted MEan RAnk gene set testing (CAMERA) of gastrocnemius
and vastus lateralis data. The gene sets available for testing, as well as
dataset-specific gene sets, are detailed in \code{\link{rat_gene_sets}}.
For metabolomics, RefMet chemical subclass sets were constructed from the
"sub_class" column in the differential analysis results. Each -omics
dataset was collapsed to the gene or RefMet metabolite/lipid ID level
(metabolomics/lipidomics only) by selecting the most extreme z-score per
contrast. See the vignettes for details.
}
\examples{
# Number of differential molecular signatures (FDR = 0.05)
f <- function(x) {
  lapply(x, function(.x) with(.x, table(Contrast, PAdj < 0.05)))
}

## FGSEA ----
# Gastrocnemius
#f1(TRNSCRPT_GN_FGSEA) # Transcriptomics
#f1(PROT_GN_FGSEA)     # Proteomics
#f1(PHOSPHO_GN_FGSEA)  # Phosphoproteomics
#f1(ACETYL_GN_FGSEA)   # Acetyl proteomics
#f1(REDOX_GN_FGSEA)    # Redox proteomics
#f1(METAB_GN_FGSEA)    # Metabolomics/lipidomics

# Vastus lateralis
#f1(TRNSCRPT_VL_FGSEA) # Transcriptomics
#f1(METAB_VL_FGSEA)    # Metabolomics/lipidomics

## CAMERA-PR ----
# Gastrocnemius
#f1(TRNSCRPT_GN_CAMERA) # Transcriptomics
#f1(PROT_GN_CAMERA)     # Proteomics
#f1(PHOSPHO_GN_CAMERA)  # Phosphoproteomics
#f1(ACETYL_GN_CAMERA)   # Acetyl proteomics
#f1(REDOX_GN_CAMERA)    # Redox proteomics
#f1(METAB_GN_CAMERA)    # Metabolomics/lipidomics

# Vastus lateralis
#f1(TRNSCRPT_VL_CAMERA) # Transcriptomics
#f1(METAB_VL_CAMERA)    # Metabolomics/lipidomics


}
\seealso{
\code{\link[fgsea]{fgsea}}, \code{\link[TMSig]{cameraPR.matrix}},
\code{\link{molecular_signatures}}
}
\keyword{datasets}
