% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SKM_REDUNDANT_SETS}
\alias{SKM_REDUNDANT_SETS}
\alias{TRNSCRPT_GN_REDUNDANT_SETS}
\alias{PROT_GN_REDUNDANT_SETS}
\alias{ACETYL_GN_REDUNDANT_SETS}
\alias{REDOX_GN_REDUNDANT_SETS}
\alias{TRNSCRPT_VL_REDUNDANT_SETS}
\alias{ACETYL_GN_NORM_REDUNDANT_SETS}
\alias{REDOX_GN_NORM_REDUNDANT_SETS}
\title{Redundant gene sets}
\format{
an object of class \code{data.frame} with 5 columns:

\describe{
\item{"GeneSetID"}{character; a unique gene set ID. Mainly used for fast
cross-referencing with the FGSEA or CAMERA-PR results.}
\item{"GeneSet"}{character; the gene set that was retained for the
analysis. Usually, the largest set in the cluster. This column does not
contain all gene sets tested—only those that appeared in a cluster with
other sets.}
\item{"SimilarGeneSets"}{character; the gene set that was excluded for analysis.}
\item{"JaccardCoef"}{numeric; the Jaccard similarity coefficient. A value
of 1 indicates that "GeneSet" and "SimilarGeneSets" are aliased, at least
after they had been restricted to only those genes measured in the
dataset.}
\item{"OverlapCoef"}{numeric; the overlap similarity coefficient. A value
of 1 indicates that "SimilarGeneSets" is a subset of "GeneSet", or they are
aliased, at least after they had been restricted to only those genes
measured in the dataset.}
}
}
\usage{
TRNSCRPT_GN_REDUNDANT_SETS
PROT_GN_REDUNDANT_SETS
ACETYL_GN_REDUNDANT_SETS
REDOX_GN_REDUNDANT_SETS

TRNSCRPT_VL_REDUNDANT_SETS
}
\description{
Gene sets that were removed following use of
\code{\link[TMSig]{clusterSets}} with default parameters. These tend to be
subsets of those sets that were retained for analysis.
}
\keyword{datasets}
